requirements.txt
torch
torchvision
numpy
pandas
albumentations
segmentation-models-pytorch
opencv-python
matplotlib
scikit-learn
Pillow
